Move input simulation in a input system subfolder and update namespace
