+++ XRTK-Core/Packages/com.xrtk.core/Services/InputSystem/Simulation/ArticulatedHandPoses/ArticulatedHandPose_PinchSteadyWrist.json.meta
