+++ XRTK-Core/Packages/com.xrtk.core/Services/InputSystem/Simulation/ArticulatedHandPoses/ArticulatedHandPose_Poke.json.meta
