+++ XRTK-Core/Packages/com.xrtk.core/Services/InputSystem/Simulation/HandSimulationMode.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6e212a75801dcf84db1cf7b01922d934
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
