+++ XRTK-Core/Packages/com.xrtk.core/Services/InputSystem/Simulation/InputSimulationControlMode.cs
@@ -1,0 +1,13 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+
+namespace XRTK.Services.InputSystem.Simulation
+{
+    public enum InputSimulationControlMode
+    {
+        // Move in the main camera forward direction
+        Fly,
+        // Move on a X/Z plane
+        Walk,
+    }
+}
