+++ XRTK-Core/Packages/com.xrtk.core/Services/InputSystem/Simulation/InputSimulationControlMode.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d6a7ca489070ec346a7272240de463e1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
