+++ XRTK-Core/Packages/com.xrtk.core/Services/InputSystem/Simulation/InputSimulationMouseButton.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
-namespace XRTK.Services.InputSimulationSystem
+namespace XRTK.Services.InputSystem.Simulation
 {
     /// <summary>
     /// This enum is used to customize how/when users will look around in the Unity player
@@ -17,23 +17,4 @@ namespace XRTK.Services.InputSimulationSystem
         Focused,    // When Unity player has focus
         None        // No mouse look functionality
     }
-
-    public enum InputSimulationControlMode
-    {
-        // Move in the main camera forward direction
-        Fly,
-        // Move on a X/Z plane
-        Walk,
-    }
-
-    public enum HandSimulationMode
-    {
-        // Disable hand simulation
-        Disabled,
-        // Raises gesture events only
-        Gestures,
-        // Provide a fully articulated hand controller
-        Articulated,
-    }
-
 }
