+++ XRTK-Core/Packages/com.xrtk.core/Services/InputSystem/Simulation/ManualCameraControl.cs.meta
