+++ XRTK-Core/Packages/com.xrtk.core/Services/InputSystem/Simulation/SimulatedArticulatedHand.cs.meta
