+++ XRTK-Core/Packages/com.xrtk.core/Services/InputSystem/Simulation/SimulatedGestureHand.cs
@@ -7,7 +7,7 @@ using XRTK.Definitions.InputSystem;
 using XRTK.Definitions.Utilities;
 using XRTK.Interfaces.InputSystem;
 
-namespace XRTK.Services.InputSimulationSystem
+namespace XRTK.Services.InputSystem.Simulation
 {
     public class SimulatedGestureHand : SimulatedHand
     {
