+++ XRTK-Core/Packages/com.xrtk.core/Services/InputSystem/Simulation/SimulatedGestureHand.cs.meta
