+++ XRTK-Core/Packages/com.xrtk.core/Services/InputSystem/Simulation/SimulatedHand.cs
@@ -8,7 +8,7 @@ using XRTK.Definitions.Utilities;
 using XRTK.Interfaces.InputSystem;
 using XRTK.Providers.Controllers.Hands;
 
-namespace XRTK.Services.InputSimulationSystem
+namespace XRTK.Services.InputSystem.Simulation
 {
     public abstract class SimulatedHand : BaseHand
     {
