+++ XRTK-Core/Packages/com.xrtk.core/Services/InputSystem/Simulation/SimulatedHand.cs.meta
