+++ XRTK-Core/Packages/com.xrtk.core/Services/InputSystem/Simulation/SimulatedHandData.cs.meta
